Properly define CFLAGS and remove some warnings.
There is a lot of more warnings with -Wall; for now
just disable the flag.

+++ lib/status.h
@@ -3,5 +3,5 @@ struct progress {
 	unsigned long current;
 };
 
-unsigned get_progress_percentage(struct progress *prog);
-unsigned inc_and_get_progress_percentage(struct progress *prog);
+unsigned long get_progress_percentage(struct progress *prog);
+unsigned long inc_and_get_progress_percentage(struct progress *prog);
